separate out models and schemas
