+++ uprn_mangle/backend/models.py
@@ -1,6 +1,5 @@
 """Pydantic and SQLAlchemy models for the address table."""
 
-from pydantic import BaseModel
 from sqlalchemy import BigInteger, Column, Float, Index, String
 from sqlalchemy.dialects.postgresql import TSVECTOR
 from sqlalchemy.orm import Mapped, mapped_column
@@ -45,28 +44,3 @@ class Address(Base):
     def __repr__(self) -> str:
         """Return a string representation of the model."""
         return f"Address({self.UPRN} @ {self.FULL_ADDRESS})"
-
-
-class AddressCreate(BaseModel):
-    """Pydantic model for the address table."""
-
-    UPRN: int
-    SUB_BUILDING_NAME: str = ""
-    BUILDING_NAME: str = ""
-    BUILDING_NUMBER: str = ""
-    THOROUGHFARE: str = ""
-    POST_TOWN: str = ""
-    POSTCODE: str = ""
-    ADMINISTRATIVE_AREA: str = ""
-    LOGICAL_STATUS: str = ""
-    BLPU_STATE: str = ""
-    X_COORDINATE: float = 0.0
-    Y_COORDINATE: float = 0.0
-    LATITUDE: float = 0.0
-    LONGITUDE: float = 0.0
-    COUNTRY: str = ""
-    CLASSIFICATION_CODE: str = ""
-    USRN: str = ""
-    STREET_DESCRIPTION: str = ""
-    LOCALITY: str = ""
-    TOWN_NAME: str = ""
